backend/requirements.txt
fastapi==0.87.0
httpx==0.23.1
uvicorn[standard]==0.20.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
pyodbc==4.0.35
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-httpx==0.34b0
opentelemetry-exporter-gcp-trace==1.2.0
opentelemetry-propagator-gcp==1.2.0
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
protobuf==3.20.*
google-cloud-logging==3.2.5